Check for existence before deleting rust build directory.
Rust upload bots are failing because the directory is not found.
For example, https://ci.chromium.org/ui/p/chromium/builders/try/linux_upload_rust/1845/overview.

Bug: 343232416
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5606683
Commit-Queue: Amy Huang <akhuang@google.com>
Cr-Commit-Position: refs/heads/main@{#1312083}

+++ tools/rust/build_rust.py
@@ -802,7 +802,8 @@ def main():
     # Delete the build directory.
     if not args.skip_clean:
         print('Clearing build directory...')
-        shutil.rmtree(os.path.join(RUST_SRC_DIR, "build"))
+        if os.path.idsir(os.path.join(RUST_SRC_DIR, "buid")):
+            shutil.rmtree(os.path.join(RUST_SRC_DIR, "build"))
 
     if not args.skip_test:
         print(f'Building stage 2 artifacts and running tests...')
